+++ TourGuide/src/test/java/com/openclassrooms/tourguide/unitTest/TestRewardsService.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.tourguide;
+package com.openclassrooms.tourguide.unitTest;
 
 import com.openclassrooms.tourguide.helper.InternalTestHelper;
 import com.openclassrooms.tourguide.service.RewardsService;
